+++ probe-rs/src/architecture/arm/core/armv8a.rs
@@ -714,6 +714,10 @@ impl<'probe> Armv8a<'probe> {
 
         Ok(())
     }
+
+    fn set_core_status(&mut self, new_status: CoreStatus) {
+        super::update_core_status(&mut self.memory, &mut self.state.current_state, new_status);
+    }
 }
 
 impl<'probe> CoreInterface for Armv8a<'probe> {
@@ -816,7 +820,7 @@ impl<'probe> CoreInterface for Armv8a<'probe> {
         }
 
         // Recompute / verify current state
-        self.state.current_state = CoreStatus::Running;
+        self.set_core_status(CoreStatus::Running);
         let _ = self.status()?;
 
         // Gate restart channel
@@ -1041,7 +1045,7 @@ impl<'probe> CoreInterface for Armv8a<'probe> {
         if edscr.halted() {
             let reason = edscr.halt_reason();
 
-            self.state.current_state = CoreStatus::Halted(reason);
+            self.set_core_status(CoreStatus::Halted(reason));
             self.state.is_64_bit = edscr.currently_64_bit();
 
             return Ok(CoreStatus::Halted(reason));
@@ -1051,7 +1055,7 @@ impl<'probe> CoreInterface for Armv8a<'probe> {
             tracing::warn!("Core is running, but we expected it to be halted");
         }
 
-        self.state.current_state = CoreStatus::Running;
+        self.set_core_status(CoreStatus::Running);
 
         Ok(CoreStatus::Running)
     }
@@ -1290,6 +1294,8 @@ mod test {
     }
 
     impl ArmProbe for MockProbe {
+        fn update_core_status(&mut self, _: CoreStatus) {}
+
         fn read_8(&mut self, _address: u64, _data: &mut [u8]) -> Result<(), ArmError> {
             todo!()
         }
@@ -1375,7 +1381,9 @@ mod test {
             >,
             DebugProbeError,
         > {
-            todo!()
+            Err(DebugProbeError::NotImplemented(
+                "get_arm_communication_interface",
+            ))
         }
 
         fn read_64(&mut self, _address: u64, _data: &mut [u64]) -> Result<(), ArmError> {