+++ probe-rs/src/architecture/arm/core/mod.rs
@@ -1052,3 +1052,23 @@ impl CortexAState {
         self.initialized
     }
 }
+
+/// Core implementations should call this function when they
+/// wish to update the [`CoreStatus`] of their core.
+///
+/// It will reflect the core status to the probe/memory interface if
+/// the status has changed, and will replace `current_status` with
+/// `new_status`.
+pub fn update_core_status<
+    P: super::memory::adi_v5_memory_interface::ArmProbe + ?Sized,
+    T: core::ops::DerefMut<Target = P>,
+>(
+    probe: &mut T,
+    current_status: &mut CoreStatus,
+    new_status: CoreStatus,
+) {
+    if *current_status != new_status {
+        probe.deref_mut().update_core_status(new_status);
+    }
+    *current_status = new_status;
+}