+++ probe-rs/src/probe/jlink/arm.rs
@@ -1386,6 +1386,10 @@ impl<Probe: DebugProbe + RawProtocolIo + JTAGAccess + 'static> RawDapAccess for 
 
         Ok(())
     }
+
+    fn core_status_notification(&mut self, _: crate::CoreStatus) -> Result<(), DebugProbeError> {
+        Ok(())
+    }
 }
 
 #[cfg(test)]