+++ CHANGELOG.md
@@ -8,10 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - probe-rs-debugger: Show errors that happen before VSCode/DAP Client session initializion has completed (#1581).
+- probe-rs-cli-util: replace unwanted instance of `println` with `eprintln` (#1595, fixes #1593).
 
 ### Added
 
 - Added support for the Olimex ARM-USB-TINY-H JTAG device (#1586).
+- Added support for propagating `CoreStatus` to the probe in use (#1588).
 
 ## [0.18.0]